requirements.txt
aiofiles==25.1.0
aiohttp==3.13.2
aiosqlite==0.22.1
async_timeout==5.0.1
beautifulsoup4==4.14.3
discord.py[speed,voice]==2.6.4
google-cloud-error-reporting==1.14.0
google-genai[aiohttp]==1.62.0
grpcio==1.76.0
html5lib==1.1
humanfriendly==10.0
mypy-extensions==1.1.0
mypy==1.19.1
parsedatetime==2.6
Pillow==12.1.0
protobuf==6.33.5
PyNaCl==1.5.0
python-dotenv==1.2.1
pytz==2025.2
RapidFuzz==3.14.3
ruff==0.14.14
git+https://github.com/Rapptz/discord-ext-menus@master # menus
# pip install -r requirements.txt --use-deprecated=legacy-resolver
# -- grpcio --
# https://github.com/grpc/grpc/issues/24249#issuecomment-972555615
# sudo find / -type f -name '*atom*.so*'
# export LD_PRELOAD=/usr/lib/arm-linux-gnueabihf/libatomic.so.1.2.0
